. (#1)
* Initial commit of the object mother #498

Add a simple king and queen classes which implement the behaviour of the
royalty interface. Also wrote the object mother of royalty objects which
is final so you can just call the static methods in it to create objects
with a specific state to use them fast in tests. The tests are already
created for testing the behaviour and the type of the objects which are
created by the object mother.
I also created the UML diagrams via object aid and updated the readme.

* Add PlantUML file for the UML diagram #498

* End process logic clause has been corrected.

* Unused import removed.

* Add SonarQube.com badge

* #502 Replaced usages of System.out with logger.

* Change name of variables in test #498

* add entry to parent pom.xml #498

* Fix pom.xml

* Fix pom.xml

Deleted </modules> tag because i added it one time more than neccessary

* FirstCut++

FirstCut++

* SecondCut++

SecondCut++

* App++

App++

* App

App

* checkstyle errors removed

checkstyle errors removed

* #502 Reverted changes in composite example, due to nature of this example using logger isn't good idea

* Remove use of coveralls-maven-plugin (sonarqube.com covers this)

* #502 Adjusted tests for logger introduction

* #502 Introduced logging in new example

* Create presentation template

* Link Hexagonal Architecture pattern to corresponding blog entry

* Work on Hexagonal Architecture presentation

* Add new diagram

* Add presentation link to README.md

* Update App.java

The relationship between decorator and target (in the code example between Troll und SmartTroll) ist not a composition as described in the code comment lines of class App. It's a aggregation. In a composition both instances canncot live without the other one. But the target (Troll) could live without the decorator (Smart Troll). Changed "composition" to "aggregation" in the comment lines. Code lines are unchanged.
Greetings Thomas

* Review comments++

* Review comments++

* #502 Reduced amount of information logged

* Issue #450. First commit. Queue Based Load Leveling

* Issue #450. Updated pom.xml

* Updated checkstyle errors.

* Updated pom.xml with maven-pmd-plugin

Added dependency "maven-pmd-plugin" to the pom file.

* Reverted pom.xml changes that are done previously.

* Fixed PMD issue.

* Decorator pattern: Improve the example

* Removed unused imports in tests

* Proxy pattern: Improve the example

* Proxy pattern: Add missing license header

* Update URM Version to 1.4.4 and regenerated all puml files

* Update license headers

* Fixed all the code changes after review

* Review comments++

* Sonar bug fixes using Yoda condition in equals expression when comparing String literal with String object.
Using try-with-resources if we use Scanner to close the underlying stream is a good practice to handle resources.
Minimal refactor.

* Sonar bug fixes using Yoda condition in equals expression when comparing String literal with String object.
Using try-with-resources if we use Scanner to close the underlying stream is a good practice to handle resources.
Minimal refactor.

* Adding initialization on demand holder idiom.

* Changing constant interface pattern with a Noninstantiable class

* Add some material for proxy presentation.

* Add presentation for Proxy pattern.

* Add link to Proxy presentation.

* Achieved milestone 1.14.0

* Set version for next development iteration

* #525 Add link to Queue-Based Load Leveling blog

* #525 Add link to proxy blog

* Changing code to use interfaces instead of implementations.

* Changing code to use interfaces instead of implementations.

* Reverting initialization on demand holder idiom.

* #190 Regenerate puml files

* #190 Rename package for Queue-Based Load Leveling pattern